@@ -9,38 +9,47 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@heroicons/react": "^2.0.18",
+    "@heroicons/react": "^2.1.1",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-tabs": "^1.0.4",
     "@radix-ui/react-toast": "^1.1.5",
     "@radix-ui/react-toggle": "^1.0.3",
     "@radix-ui/react-toggle-group": "^1.0.4",
+    "@t3-oss/env-core": "^0.7.1",
     "@vercel/analytics": "^1.1.1",
     "class-variance-authority": "^0.7.0",
-    "clsx": "^2.0.0",
+    "clsx": "^2.1.0",
     "contentlayer": "^0.3.4",
     "date-fns": "^2.30.0",
-    "framer-motion": "^10.16.5",
+    "framer-motion": "^10.16.16",
     "lucide-react": "^0.294.0",
+    "mdast-util-toc": "^7.0.0",
     "next": "14.0.3",
     "next-contentlayer": "^0.3.4",
     "next-themes": "^0.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "tailwind-merge": "^2.1.0",
-    "tailwindcss-animate": "^1.0.7"
+    "remark": "^15.0.1",
+    "tailwind-merge": "^2.2.0",
+    "tailwindcss-animate": "^1.0.7",
+    "unist-util-visit-parents": "^6.0.1",
+    "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "autoprefixer": "^10.0.1",
-    "eslint": "^8",
+    "@types/node": "^20.10.6",
+    "@types/react": "^18.2.46",
+    "@types/react-dom": "^18.2.18",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8.56.0",
     "eslint-config-next": "14.0.3",
-    "postcss": "^8",
+    "postcss": "^8.4.32",
     "reading-time": "^1.5.0",
-    "tailwindcss": "^3.3.0",
-    "typescript": "^5"
+    "rehype-autolink-headings": "^7.1.0",
+    "rehype-pretty-code": "^0.12.3",
+    "rehype-slug": "^6.0.0",
+    "remark-gfm": "^3.0.1",
+    "tailwindcss": "^3.4.0",
+    "typescript": "^5.3.3"
   }
 }
